+++ source/extensions/omni.flux.nvtt.core/omni/flux/nvtt/core/tests/unit/__init__.py
@@ -14,16 +14,3 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 """
-
-__all__ = (
-    "DDS_CONVERSION_SETTINGS_METADATA_KEY",
-    "DDS_SOURCE_HASH_METADATA_KEY",
-    "DDS_TEXTURE_TYPE_METADATA_KEY",
-    "ORPHAN_PARAMETER_CLEANUP_SETTING_PATH",
-)
-
-DDS_SOURCE_HASH_METADATA_KEY = "asset_pipeline_source_hash"
-DDS_TEXTURE_TYPE_METADATA_KEY = "asset_pipeline_texture_type"
-DDS_CONVERSION_SETTINGS_METADATA_KEY = "asset_pipeline_conversion_settings"
-
-ORPHAN_PARAMETER_CLEANUP_SETTING_PATH = "/exts/omni.usd/mdl/ignoreOrphanParametersCleanup"